Fix range update to legacy2 with clone
When a new clone is deployed it get the initial allocation from the
current range or the next range already allocated. The code was not
considering the case of next range so it could generate a range overlap.

The fix will check if the next range is totally allocate for the service
or it is partial and in this case only the remaining part is used.