[RHCS-5445] Implement new SSN legacy2 generator for CA
The legacy id generator can introduce gaps when new ranges are allocated
because of some conversion errors between hex and decimal. A new serial id
generator has been introduced to avoid gap problems.

The old and new serial generator are described in the following pages:

- SSNv1: https://github.com/dogtagpki/pki/wiki/Sequential-Serial-Numbers-v1
- SSNv2://github.com/dogtagpki/pki/wiki/Sequential-Serial-Numbers-v2

Instances using SSNv1  can migrate to the SSNv2 using the commands
described here:

- https://github.com/dogtagpki/pki/wiki/Migrating-to-Sequential-Serial-Numbers-v2

If there is no need to avoid gaps it is better to avoid the migration.

NOTE: the links above provide documentation to the newer version of
dogtag. The porting to this release has some limitations to consider:

- the new generator is available only to the CA ids;
- some of the parameter are not customisable with `pkispawn` (e.g. increment, transfer, etc..) only the initial range limits.

Additionally, during the migration of a clone it is required to provide
new range limits even they will be overwritten with values coming from
master.