Fix mail notification CI
The notification was not working because of a CVE in postfix required to
modify the default policy becoming more restrictive.

The fix is documented here: https://bugzilla.redhat.com/show_bug.cgi?id=2255563

This is a workaround which put back the original policy. The proper fix
would be to update the code communicating with smtp server.