Fix logrotate error
If the log rotate is invoked following a log using slf4j-impl the current
policy denied the accees to the log folder and the rotate fails.

This is always the case with the change to the SessionTimer class.

To solve the problem the policy is updated to grant slf4j-impl
read/write access to log folders.