Update pki CLI to use ArgumentParser
The pki CLI has been updated to parse the main arguments (which
are also defined in MainCLI.java) using ArgumentParser, determine
the subcommand, then pass the remaining arguments to the Python
or Java module corresponding to the subcommand.

It does not use subparsers like in pki-server since the pki CLI
does not have the list of Java subcommands.

The pki CLI have options that might conflict with other options
in some subcommands. That will be addressed separately later.

The help message for default message format in MainCLI.java has
also been updated.