Fix CLI help messages
Previously if a user called pki-server CLI with a wrong
sub-command the ArgumentParser would show an auto-generated
error message which looks significantly different from the
help message already defined in print_help().

To fix the issue the PKIServerCLI.create_parser() and
execute() have been modified to use remainder instead of
subparsers, and the create_parser() in the sub-commands has
been modified to create a regular ArgumentParser instead of
a subparser.

Similar changes were made to the CLI base class to fix the
help messages for sub-commands (e.g. pki-server ca).

The pki-server has also been modified to provide a --version
option to show the version number of the tool.

A new CLIException has been added to distinguish a normal
CLI error (which will generate a simple error message such
as "Invalid module") from an unexpected exception which will
generate a full stack trace.

Resolves: https://github.com/dogtagpki/pki/issues/4932