Add test for installing CA with obsolete admin cert
The test for installing CA with existing config has been updated
to reinstall the CA with new system certs but with an old admin
cert from the previous installation (which should fail). Then the
test will remove the admin cert and reinstall the CA again (which
should succeed) and a new admin cert should be created.

pkispawn was updated to indicate that the installation failure
might be caused by an obsolete admin cert.