Don't use pki_token_name as default token
Previously the pki_token_name was used as the default token for
all system certs, but due to a bug the fallback mechanism has not
been working properly in all cases, so currently the token name
for each system cert has to be specified in pki_<cert>_token if
it's needed, which is actually simpler.

Since the bug has existed for a long time and might require a lot
of effort to properly fix and test it, the code that provides the
fallback mechanism is no longer needed and has been dropped.