Don't create audit signing cert by default
Previously the audit signing cert was always created by default
during installation, but it's actually only used if log signing
is enabled.

To reduce redundancies the default pki_audit_signing_nickname
has been removed so that pkispawn will skip creating the audit
signing cert if the nickname is not specified.

The selftest and pki-healthcheck have been modified to skip
validating the audit signing cert if the nickname is not
available.

The pki-server cert-find and pki-server subsystem-cert-find
commands have also been modified to skip displaying the audit
signing cert if the nickname is not available.

The basic tests of each subsystem have been updated to install
the subsystem without the audit signing cert nickname so it will
not create the cert. For basic CA test, it will later create the
cert and enable log signing manually.