Fix missing error messages
Previously when a failure happened in the Python code in some
cases the error message was missing which made it difficult to
troubleshoot the problem.

To simplify troubleshooting some of the Python code has been
updated to capture only the stdout and let the stderr appear
on the console so that it can be seen by the user.

The PKISubsystem.run() has been modified to provide stdout and
stderr params so they can be configured separately.