Refactor AuthorityRepository.findCAs()
Currently the AuthorityRepository.findCAs() gets the authority
data from the CA objects in the CA engine. The AuthorityMonitor
creates a CA object for each authority record in LDAP, so it
might not scale well with large database and might be expensive
to run in cloud.

To address the issue, the AuthorityRepository.findCAs() has been
modified to get the data from the authority records directly. For
now it still uses the CA object to get some additional info, but
in the future it might be possible to store those info in LDAP as
well.

The AuthorityService.findCAs() has been modified to use the
AuthorityRepository.findCAs().

The AuthorityServlet.findCAs() has been modified to no longer
catch the exception and throw a BadRequestException since the
exception might not necessarily be a client issue.