Move AuthorityMonitor.authorities into CAEngine
The authorities map in AuthorityMonitor has been moved into
CAEngine and can only be accessed through synchronized methods
to avoid race conditions. Later there might be additional fields
that will be moved from AuthorityMonitor to CAEngine as well.

The CAEngine.getCA() has been modified to look for the requested
CA object in the map. If the CA object doesn't exist it will be
loaded from the database and stored in the map.

The code in AuthorityMonitor.addCA() and remove() that updates
the authorities map has been moved into CAEngine as well.