Disable password console callback if file provided
When a password file is provided to the CLI then all password are read
from that file and no requests are done to the user if a token misses
password.

Missing password tokens are reported in verbose or debug mode.

The console password callback is still used if the user provides
password for individual tokens.

Fix issue #5045.