Update standalone OCSP test
The test for standalone OCSP has been updated to install the
OCSP without a security domain while the CA is down. The CRL
publishing will be set up after installation.

In the future all OCSP installations might be done this way
to reduce the complexity of the installation code and to
make it more reliable (i.e. less dependent on a running CA).

The pki-server status has also been updated to support OCSP
that does not have any security domain params.