Fix pki CLI help commands list
A list of java commands is included in the python code to be included in
the help.

The pki help command, or when no command is provided, returns a list of
commands based on the initialised python CLI modules ignoring the commands
implemented by the java CLI. The new list include these and it is
printed in the help so all possible commands are shown.

Fixes: #5061