Add a new profile input for client information
Implement a new profile input `ClientAuthInfoInput` for information
about client authentication.

When a RA send an enrollment request to the CA the new input allows to
provide information about the client originating the request in order to
implement ad-hoc policy. This could be relevant for some enrollment
such as EST.