Add a pattern params to RAClientAuthSubjectNameContraint
The new params allows to define subject with additional information.
The pattern param has to match with the new subject after replacing the
2 possible values with the actual values:

- $ra_client_name$
- $ra_client_uid$

It is not handled like a regular expression.

The pattern is only used when client certificate is not available.

Finally, the estServiceCert.cfg has been modified to include the new
constraint (enabled) as well as the SubjectNameContraint so the user can
easily modify the policy to apply.