Update TPS container test to use generic subsystem users
Normally pkispawn would create subsystem users with a username
that contains the hostname and the port number of the subsystem
(e.g. CA-ca.example.com-8443). However, if the subsystem is
migrated to a different server the username will no longer be
correct and there's no official process to fix it. Also, in a
cloud environment the hostnames of the containers may be random
and not permanent.

To avoid issues or confusions it's recommended to use generic
subsystem users without the hostname or port number in the
username (e.g. CA). The TPS container test has been updated to
use generic subsystem users to demonstrate that it's already
supported by the current code.

In the future pkispawn might be modified to provide params to
change the usernames of the subsystem users.