Fix nuxwdog instances
Nuxwdog enabled instances are not working with latest release of systemd
for permission problems.

With the release of systemd version 257 the command
`systemd-ask-password` when executed as user in systemd servive cannot invoke agent for
password request because it is missing privileged. At the same time
cannot execute as system because pkiuser is not authorised.

The TTY agent alternative is not working properly because it is not
linked to the root shell executing the start command.

As a solution the `pki-server-nuxwdog` script is executed with privileged
user but it operates on user keyring. This required a new option with
user name and the possibility for keyring object to operate for a
specific user.

Fix #5085