Add internationalization test
The PKIClient.entity() and unmarshall() has been updated
to create the request object and to parse the response
object with UTF-8 encoding.

The SimpleProperties.load() and store() has been updated
to load and to store the profile config file with UTF-8
encoding.

The ConfigStore.store() has been updated to export the
profile config into a raw format with UTF-8 encoding.

The pki <subsystem>-user-add has been modified to accept
a --full-name option for consistency with the pki-server
<subsystem>-user-add command.

The UserShowCLI has been modified to retrieve the correct
fields from the JSON object.

A new test has been added to create a cert profile, enroll
a cert, create a user, then perform client cert auth with
wide characters.