Revert OMisc fixes provided to cope with last minute tomcat 10.1.36 surprises. (#5153)"
This reverts commit b58e282a357e7664526c38db8c95d2f6089e916f.

This essentially allows pki to run under a tomcat that behaves like
tomcat-10.1-34. This is in anticipation of a new version 36 tomcat
build reverted to version 34 behavior.

Update requires to tomcat-10.1-43, which has been rebuilt to version 34 contents.
Redo tests after adjusting copr tomcat build to the latest.

Update to tomcat-10-1-43-5.