Replace java_devel dependency in pki-server
pki-server currently requires java_devel but it is not necessary for the
server to work properly.

Since this requirement has many dependencies to desktop related packages it
is replaced with the java_headless dependency which is enough to work.

The only tool used from java_devel is the command jar which has been
replaced by unzip since they support the same format.