Modify EST deployment to match other subsystems
EST subsystem has been modified to use the security domain during the
deployment and this is the default behaviour.

Installing with SD the EST subsystem will use the subsystem certificate
to communicate with the CA. An additional subsystem user is created when
deployed in a separate instance.

Alternatively, standalone 2 step installation is provided and in this
case the sslserver and subsystem certificate have to be generated during
installation.

EST cannot create its own SD since the related APIs are not present.