Fix update with nuxwdog
Upgrade/migrate script trying to read certificate information from
NSSDB read all the passwords from the related file but if it is not present
then only the password for the current token is read from keyctl. As a
result, the token has to be specified when DB is open or the operation
will fail.

An example of upgrade script accessing certificate information is in
base/server/upgrade/11.5.0/04-RemoveCertCSRfromConfig.py

Fix #4895