Add CLI option to specify DS backend
The pki-server acme-database-init and acme-database-index-rebuild
have been modified to provide an option to specify the DS backend
for creating/rebuilding indexes.

The test for ACME with separate instance has been modified to use
a non-default DS backend.

Note that pkispawn will not be modified to initialize the database
(which is the current behavior when installing ACME with pkispawn).
Instead, for installing the first ACME instance the admin will need
to initialize the database using the above commands after running
pkispawn, and for installing an ACME replica the admin will need to
configure the DS replication separately.

https://github.com/dogtagpki/pki/wiki/PKI-Server-ACME-Database-CLI
https://github.com/dogtagpki/pki/wiki/PKI-Server-ACME-Realm-CLI