IDM-3425 Complete CI and merge for dogtag-pki changes for f43 to master upstream branch.
This fix allows pki to build and run on either a tomcat 9 or tomcat 10 system.
If pki is built on a tomcat 9 system, it will run on a tomcat 9 platform.
If pki is built on a tomat 10 system, it will run on a tomcat 10 system.

Tomcat 10 is used for platforms >= f43 and >= rhel10.

Fix build.sh to determine the APP_SERVER variable without obtaining it from the spec file.

Account for jdk versions for rhel, different from fedora.

Address review comments.