Add key ID file option
The pki nss-key-create, nss-key-show, and nss-cert-request
commands have been modified to provide an option to specify
a file to store the ID of a new key or to load the ID of an
existing key.

The test for PKI CLI with ECC has been updated to use the
new option.

https://github.com/dogtagpki/pki/wiki/PKI-NSS-Key-CLI
https://github.com/dogtagpki/pki/wiki/Generating-Certificate-Request-with-PKI-NSS