Fix keyring exception for missing id
When a password is searched in keyring but it is not present the
following log message fails with a TypeError because the key id is null.
This error is not handled and the startup fails with a message which
make not clear what is generating the problem.

The behaviour is modified to raise a value error which is handled and if
the password cannot be retrieved the final error will be something like:

Exception: No available password to access the token "internal"