Store and use the successful api_path in InfoClient.get_info()
This is the only place where the remote available API
can be detected. This value is used by PKIClient to
to determine which endpoint to use rather than only
relying on the local package version to do so.

This adds a fake ApiPath value to the returned JSON which
in turn will add it to the object when parsed to try to
remain consistent.

Signed-off-by: Rob Crittenden <rcritten@redhat.com>