Remove check for pki log folder
Remove the check to pki logs folder because it is not always present.

Tomcat 10.1 [1] has modified the logging system introducing a lazy creation
so log folder and file are created only if there is something to write.
Therefore, the pki folder for pki webapps logs is not created since at
default logging level there should be no logs.

1. https://fedoraproject.org/wiki/Changes/Tomcat10ChangeProposal#Detailed_Description