Get dogtag working under old tomcat 10 or new tomcat 10.
New code to reconcile the proper systemd unit to use in old vs new tomcat 10 scenario, assisted by coding assitant.

The fix to the unit files uses a script that detects if we have upgraded tomcat from old tomcat 10 to new tomcat10.
If that is the case it replaces the current unit file with a newer copy of the unit file that supports new tomcat10.

This solution was deemed the most simple after several alternate attempts to get this working. Also this supports both
fresh installs under old or new tomcat 10 and also supports an independent upgrade from old to new tomcat 10.
At the next restart the unit file will be updated.

Based upon further feedback, I"ve implemented a solution based using one single file but having it call abstracted
scripts to start and stop the server, which determines which tomcat 10 is present and calls the correct start script.